Move getOffenceCode to root lib folder
