+++ packages/core/comparison/lib/hoOffencesAreEqual.ts
@@ -1,7 +1,7 @@
 import type { Offence } from "../../types/AnnotatedHearingOutcome"
 
+import getOffenceCode from "../../lib/getOffenceCode"
 import offencesHaveEqualResults from "../../phase1/enrichAho/enrichFunctions/matchOffencesToPnc/offencesHaveEqualResults"
-import getOffenceCode from "../../phase1/lib/offence/getOffenceCode"
 
 const hoOffencesAreEqual = (offence1: Offence, offence2: Offence): boolean =>
   getOffenceCode(offence1) === getOffenceCode(offence2) && offencesHaveEqualResults([offence1, offence2])
