+++ packages/core/lib/getOffenceCode.ts
@@ -1,4 +1,4 @@
-import type { Offence } from "../../../types/AnnotatedHearingOutcome"
+import type { Offence } from "../types/AnnotatedHearingOutcome"
 
 const getOffenceCode = (offence: Offence): string | undefined => {
   const offenceReason = offence.CriminalProsecutionReference.OffenceReason
