+++ packages/core/phase1/enrichAho/enrichFunctions/matchOffencesToPnc/generateCandidate.ts
@@ -2,8 +2,8 @@ import type { Offence } from "../../../../types/AnnotatedHearingOutcome"
 import type { PncOffenceWithCaseRef } from "../../../enrichAho/enrichFunctions/matchOffencesToPnc/matchOffencesToPnc"
 import type { Candidate } from "../../../enrichAho/enrichFunctions/matchOffencesToPnc/OffenceMatcher"
 
+import getOffenceCode from "../../../../lib/getOffenceCode"
 import offenceIsBreach from "../../../enrichAho/enrichFunctions/matchOffencesToPnc/offenceIsBreach"
-import getOffenceCode from "../../../lib/offence/getOffenceCode"
 import { datesMatchApproximately } from "./datesMatchApproximately"
 import { datesMatchExactly } from "./datesMatchExactly"
 import { offenceManuallyMatches } from "./offenceManuallyMatches"
