+++ packages/core/phase1/lib/isReopenedOrStatutoryDeclarationCase.ts
@@ -1,6 +1,6 @@
 import type { AnnotatedHearingOutcome } from "../../types/AnnotatedHearingOutcome"
 
-import getOffenceCode from "../lib/offence/getOffenceCode"
+import getOffenceCode from "../../lib/getOffenceCode"
 
 const matchingCodes = ["MC80524", "MC80527", "MC80528"]
