+++ packages/core/phase1/scripts/extendedSummary.ts
@@ -8,10 +8,10 @@ import type { Trigger } from "../../types/Trigger"
 
 import getFile from "../../comparison/lib/getFile"
 import { lookupOffenceByCjsCode, lookupResultCodeByCjsCode, lookupVerdictByCjsCode } from "../../lib/dataLookup"
+import getOffenceCode from "../../lib/getOffenceCode"
 import { parseAhoXml } from "../../lib/parse/parseAhoXml"
 import parseSpiResult from "../../lib/parse/parseSpiResult"
 import transformSpiToAho from "../../lib/parse/transformSpiToAho"
-import getOffenceCode from "../lib/offence/getOffenceCode"
 
 interface ComparisonFile {
   annotatedHearingOutcome: string
