+++ packages/core/phase3/lib/hearingDetails/createCourtHearingFromOffence.ts
@@ -1,7 +1,7 @@
 import type { Offence } from "../../../types/AnnotatedHearingOutcome"
 import type { CourtHearing } from "../../types/HearingDetails"
 
-import getOffenceCode from "../../../phase1/lib/offence/getOffenceCode"
+import getOffenceCode from "../../../lib/getOffenceCode"
 import { HearingDetailsType } from "../../types/HearingDetails"
 import { preProcessOffenceReasonSequence } from "./preProcessOffenceReasonSequence"
